Distributed pipeline flow tests across all pipeline stages
